+++ spf.go
@@ -12,184 +12,6 @@ import (
 	"strings"
 )
 
-/*
-* Mechanism struct and associated methods.
- */
-
-// Mechanism represents a single mechanism in an SPF record.
-type Mechanism struct {
-	Name   string
-	Domain string
-	Prefix string
-	Result string
-}
-
-// Return a Mechanism as a string
-func (m *Mechanism) String() string {
-	var buf bytes.Buffer
-
-	buf.WriteString(m.Name)
-
-	if len(m.Domain) != 0 {
-		buf.WriteString(fmt.Sprintf(":%s", m.Domain))
-	}
-
-	if len(m.Prefix) != 0 {
-		buf.WriteString(fmt.Sprintf("/%s", m.Prefix))
-	}
-
-	buf.WriteString(fmt.Sprintf(" - %s", m.Result))
-
-	return buf.String()
-}
-
-// ResultTag maps the Result code to a suitable char
-func (m *Mechanism) ResultTag() string {
-	switch m.Result {
-	case "Fail":
-		return "-"
-	case "SoftFail":
-		return "~"
-	case "Pass":
-		return "+"
-	case "Neutral":
-		return "?"
-	}
-
-	return "+"
-}
-
-// SPFString return a string representation of a mechanism, suitable for using
-// in a TXT record.
-func (m *Mechanism) SPFString() string {
-	var buf bytes.Buffer
-
-	tag := m.ResultTag()
-
-	if m.Name != "all" && tag != "+" {
-		buf.WriteString(tag)
-	} else if m.Name == "all" {
-		buf.WriteString(tag)
-	}
-
-	buf.WriteString(m.Name)
-
-	if len(m.Domain) != 0 && m.Name != "all" {
-		buf.WriteString(fmt.Sprintf(":%s", m.Domain))
-	}
-
-	if len(m.Prefix) != 0 {
-		buf.WriteString(fmt.Sprintf("/%s", m.Prefix))
-	}
-
-	return buf.String()
-}
-
-// Ensure the mechanism is valid
-func (m *Mechanism) Valid() bool {
-	var result bool
-	var name bool
-	var ip bool
-
-	switch m.Result {
-	case "Pass", "Fail", "SoftFail", "Neutral":
-		result = true
-	default:
-		result = false
-	}
-
-	switch m.Name {
-	case "all", "a", "mx", "ip4", "ip6", "exists", "include", "ptr":
-		name = true
-	default:
-		name = false
-	}
-
-	ip = true
-	if m.Name == "ip4" || m.Name == "ip6" {
-		valid := net.ParseIP(m.Domain)
-		ip = (valid != nil)
-	}
-
-	return result && name && ip
-}
-
-// Evaluate determines if the given IP address is covered by the mechanism.
-// If the IP is covered, the mechanism result is returned and error is nil.
-// If the IP is not covered an error is returned. The caller must check for
-// the error to determine if the result is valid.
-func (m *Mechanism) Evaluate(client string) (string, error) {
-
-	clientIP := net.ParseIP(client)
-
-	switch m.Name {
-	case "all":
-		return m.Result, nil
-	case "exists":
-		_, err := net.LookupHost(m.Domain)
-		if err == nil {
-			return m.Result, nil
-		}
-	case "include":
-		email := "info@" + m.Domain
-		return SPFTest(client, email)
-	case "a":
-		networks := aNetworks(m)
-		if ipInNetworks(clientIP, networks) {
-			return m.Result, nil
-		}
-	case "mx":
-		networks := mxNetworks(m)
-		if ipInNetworks(clientIP, networks) {
-			return m.Result, nil
-		}
-	case "ptr":
-		if testPTR(m, client) {
-			return m.Result, nil
-		}
-	default:
-		network, err := networkCIDR(m.Domain, m.Prefix)
-		if err == nil {
-			if network.Contains(clientIP) {
-				return m.Result, nil
-			}
-		}
-	}
-
-	return "", errors.New("Client was not covered by the mechanism.")
-}
-
-// NewMechanism creates a new Mechanism struct using the given string and
-// domain name. When the mechanism does not define the domain, the provided
-// domain is used as the default.
-func NewMechanism(str, domain string) *Mechanism {
-	m := new(Mechanism)
-
-	switch string(str[0]) {
-	case "-":
-		m.Result = "Fail"
-		parseMechanism(str[1:], domain, m)
-	case "~":
-		m.Result = "SoftFail"
-		parseMechanism(str[1:], domain, m)
-	case "+":
-		m.Result = "Pass"
-		parseMechanism(str[1:], domain, m)
-	case "?":
-		m.Result = "Neutral"
-		parseMechanism(str[1:], domain, m)
-	default:
-		m.Result = "Pass"
-		parseMechanism(str, domain, m)
-	}
-
-	return m
-}
-
-/*
- SPF Struct and associated methods.
-*/
-
 // SPF represents an SPF record for a particular Domain. The SPF record
 // holds all of the Allow, Deny, and Neutral mechanisms.
 type SPF struct {
@@ -203,7 +25,7 @@ type SPF struct {
 // Mechanisms are evaluated in order until one of them provides a valid
 // result. If no valid results are provided, the default result of "Neutral"
 // is returned.
-func (s *SPF) Test(client string) string {
+func (s *SPF) Test(client string) ResultType {
 	for _, m := range s.Mechanisms {
 		result, err := m.Evaluate(client)
 		if err == nil {
@@ -211,7 +33,7 @@ func (s *SPF) Test(client string) string {
 		}
 	}
 
-	return "Neutral"
+	return Neutral
 }
 
 // Return an SPF record as a string.
@@ -287,7 +109,7 @@ Exported functions.
 //
 // SPFTest will return one of the following results:
 // Pass, Fail, SoftFail, Neutral, None, TempError, or PermError
-func SPFTest(client, email string) (string, error) {
+func SPFTest(client, email string) (ResultType, error) {
 	var domain string
 	var spfText string
 
@@ -296,13 +118,13 @@ func SPFTest(client, email string) (string, error) {
 		parts := strings.Split(email, "@")
 		domain = parts[1]
 	} else {
-		return "", errors.New("Email address must contain an @ sign.")
+		return None, errors.New("Email address must contain an @ sign.")
 	}
 
 	// DNS errors during domain name lookup should result in "TempError".
 	records, err := net.LookupTXT(domain)
 	if err != nil {
-		return "TempError", err
+		return TempError, err
 	}
 
 	// Find the SPF record among the TXT records for the domain.
@@ -315,114 +137,14 @@ func SPFTest(client, email string) (string, error) {
 
 	// No SPF record should result in None.
 	if spfText == "" {
-		return "None", nil
+		return None, nil
 	}
 
 	// Create a new SPF struct
 	spf, err := NewSPF(domain, spfText)
 	if err != nil {
-		return "PermError", err
+		return PermError, err
 	}
 
 	return spf.Test(client), nil
-}
-
-/*
-Unexported supporting functions.
-*/
-
-func parseMechanism(str, domain string, m *Mechanism) {
-	ci := strings.Index(str, ":")
-	pi := strings.Index(str, "/")
-
-	switch {
-	case ci != -1 && pi != -1: // name:domain/prefix
-		m.Name = str[:ci]
-		m.Domain = str[ci+1 : pi]
-		m.Prefix = str[pi+1:]
-	case ci != -1: // name:domain
-		m.Name = str[:ci]
-		m.Domain = str[ci+1:]
-	case pi != -1: // name/prefix
-		m.Name = str[:pi]
-		m.Domain = domain
-		m.Prefix = str[pi+1:]
-	default: // name
-		m.Name = str
-		m.Domain = domain
-	}
-}
-
-func networkCIDR(addr, prefix string) (*net.IPNet, error) {
-	if prefix == "" {
-		ip := net.ParseIP(addr)
-
-		prefix = "32"
-		if len(ip) == 8 {
-			prefix = "128"
-		}
-	}
-
-	cidrStr := fmt.Sprintf("%s/%s", addr, prefix)
-
-	_, network, err := net.ParseCIDR(cidrStr)
-	return network, err
-}
-
-func ipInNetworks(client net.IP, networks []*net.IPNet) bool {
-	for _, network := range networks {
-		if network.Contains(client) {
-			return true
-		}
-	}
-
-	return false
-}
-
-func buildNetworks(ips []string, prefix string) []*net.IPNet {
-	var networks []*net.IPNet
-
-	for _, ip := range ips {
-		network, err := networkCIDR(ip, prefix)
-		if err == nil {
-			networks = append(networks, network)
-		}
-	}
-
-	return networks
-}
-
-func aNetworks(m *Mechanism) []*net.IPNet {
-	ips, _ := net.LookupHost(m.Domain)
-
-	return buildNetworks(ips, m.Prefix)
-}
-
-func mxNetworks(m *Mechanism) []*net.IPNet {
-	var networks []*net.IPNet
-
-	mxs, _ := net.LookupMX(m.Domain)
-
-	for _, mx := range mxs {
-		ips, _ := net.LookupHost(mx.Host)
-		networks = append(networks, buildNetworks(ips, m.Prefix)...)
-	}
-
-	return networks
-}
-
-func testPTR(m *Mechanism, client string) bool {
-	names, err := net.LookupAddr(client)
-
-	if err != nil {
-		return false
-	}
-
-	for _, name := range names {
-		if strings.HasSuffix(name, m.Domain) {
-			return true
-		}
-	}
-
-	return false
 }
